requirements.txt
certifi==2020.4.5.2
chardet==3.0.4
future==0.18.2
grpcio==1.29.0
grpcio-tools==1.29.0
idna==2.9
kafka-python==2.0.1
numpy==1.18.5
opencv-python==4.2.0.34
Pillow==7.1.2
protobuf==3.12.2
pymilvus==0.2.13
redis==3.5.3
requests==2.23.0
six==1.15.0
torch==1.5.0+cu101
torchvision==0.6.0+cu101
tqdm==4.46.1
ujson==3.0.0
urllib3==1.25.9
Flask==1.1.2
websockets==8.1
